requirements.txt
python-dotenv
spotipy
PyQt5
python-vlc
yt-dlp[default] @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz
janome
scikit-learn
platformdirs
